+++ backend/src/main/java/com/bd2_team6/biteright/controllers/requests/create_requests/UserExerciseCreateRequest.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UserExerciseCreateRequest {
-    private int exerciseInfoId;
+    private Long exerciseInfoId;
     private LocalDateTime activityDate;
     private int duration;
 }
