+++ backend/src/main/java/com/bd2_team6/biteright/controllers/requests/update_requests/MealUpdateRequest.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 public class MealUpdateRequest {
-    private Integer mealTypeId;
+    private Long mealTypeId;
     private String name;
     private String description;
     private List<MealContentDTO> contents;
