+++ backend/src/main/java/com/bd2_team6/biteright/entities/daily_limits/DailyLimitsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface DailyLimitsRepository extends JpaRepository<DailyLimits, Integer> {
+public interface DailyLimitsRepository extends JpaRepository<DailyLimits, Long> {
     
 }
