+++ backend/src/main/java/com/bd2_team6/biteright/entities/ingredient/Ingredient.java
@@ -19,7 +19,7 @@ public class Ingredient {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ingredient_id")
-    private Integer ingredientId;
+    private Long ingredientId;
     
     @Column(name = "name")
     private String name;
