+++ backend/src/main/java/com/bd2_team6/biteright/entities/limit_history/LimitHistory.java
@@ -19,7 +19,7 @@ public class LimitHistory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "history_id")
-    private Integer historyId;
+    private Long historyId;
     
     @Column(name = "date_changed")
     private LocalDate dateChanged;
