+++ backend/src/main/java/com/bd2_team6/biteright/entities/limit_history/LimitHistoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface LimitHistoryRepository extends JpaRepository<LimitHistory, Integer> {
+public interface LimitHistoryRepository extends JpaRepository<LimitHistory, Long> {
 
 }
