+++ backend/src/main/java/com/bd2_team6/biteright/entities/meal/Meal.java
@@ -10,7 +10,6 @@ import lombok.Setter;
 
 import java.time.LocalDateTime;
 import java.util.HashSet;
-import java.util.Objects;
 import java.util.Set;
 
 @Entity
@@ -22,7 +21,7 @@ public class Meal {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "meal_id")
-    private Integer mealId;
+    private Long mealId;
 
     @Column(name = "meal_date")
     private LocalDateTime mealDate;
