+++ backend/src/main/java/com/bd2_team6/biteright/entities/meal_content/MealContentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface MealContentRepository extends JpaRepository<MealContent, Integer> {
+public interface MealContentRepository extends JpaRepository<MealContent, Long> {
     
 }
