+++ backend/src/main/java/com/bd2_team6/biteright/entities/meal_type/MealType.java
@@ -18,7 +18,7 @@ public class MealType {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "meal_type_id")
-    private Integer typeId;
+    private Long typeId;
 
     @Column(name = "name")
     private String name;
