+++ backend/src/main/java/com/bd2_team6/biteright/entities/meal_type/MealTypeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface MealTypeRepository extends JpaRepository<MealType, Integer> {
+public interface MealTypeRepository extends JpaRepository<MealType, Long> {
     Optional<MealType> findByName(String name);
 }
