+++ backend/src/main/java/com/bd2_team6/biteright/entities/recipe_content/RecipeContentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface RecipeContentRepository extends JpaRepository<RecipeContent, Integer> {
+public interface RecipeContentRepository extends JpaRepository<RecipeContent, Long> {
     
 }
