+++ backend/src/main/java/com/bd2_team6/biteright/entities/user_goal/UserGoalRepository.java
@@ -1,10 +1,10 @@
 package com.bd2_team6.biteright.entities.user_goal;
+
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface UserGoalRepository extends JpaRepository<UserGoal, Integer> {
+public interface UserGoalRepository extends JpaRepository<UserGoal, Long> {
     
 }
