+++ backend/src/main/java/com/bd2_team6/biteright/entities/user_info/UserInfoRepository.java
@@ -1,10 +1,10 @@
 package com.bd2_team6.biteright.entities.user_info;
+
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface UserInfoRepository extends JpaRepository<UserInfo, Integer> {
+public interface UserInfoRepository extends JpaRepository<UserInfo, Long> {
     
 }
